Updatev3 (#150)
* pull master

* update blogs,refund, notification when transfer

+++ src/main/java/com/example/holidayswap/domain/dto/request/payment/TopUpWalletDTO.java
@@ -12,7 +12,7 @@ import lombok.Setter;
 public class TopUpWalletDTO {
     private String bankCode;
     private String orderInfor;
-    private int amount;
+    private Double amount;
     private String userId;
     private String paymentDate;
